+++ src/main/java/net/sodiumstudio/dwmg/client/DwmgClientSetupEvents.java
@@ -11,6 +11,7 @@ import com.github.mechalopa.hmag.client.renderer.KoboldRenderer;
 import com.github.mechalopa.hmag.client.renderer.MagicBulletRenderer;
 import com.github.mechalopa.hmag.client.renderer.NecroticReaperRenderer;
 import com.github.mechalopa.hmag.client.renderer.SkeletonGirlRenderer;
+import com.github.mechalopa.hmag.client.renderer.SlimeGirlRenderer;
 import com.github.mechalopa.hmag.client.renderer.SnowCanineRenderer;
 import com.github.mechalopa.hmag.client.renderer.StrayGirlRenderer;
 import com.github.mechalopa.hmag.client.renderer.WitherSkeletonGirlRenderer;
@@ -19,6 +20,7 @@ import com.github.mechalopa.hmag.client.renderer.ZombieGirlRenderer;
 import net.minecraft.client.renderer.entity.ThrownItemRenderer;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.EntityRenderersEvent;
+import net.minecraftforge.client.event.RegisterParticleProvidersEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
@@ -35,6 +37,8 @@ import net.sodiumstudio.dwmg.client.gui.screens.GuiHandItemsTwoBaubles;
 import net.sodiumstudio.dwmg.client.gui.screens.GuiImp;
 import net.sodiumstudio.dwmg.client.gui.screens.GuiKobold;
 import net.sodiumstudio.dwmg.client.gui.screens.GuiNecroticReaper;
+import net.sodiumstudio.dwmg.client.gui.screens.GuiSlimeGirl;
+import net.sodiumstudio.dwmg.client.particles.MagicalGelBallParticle;
 import net.sodiumstudio.dwmg.client.renderer.BefriendedCreeperGirlRenderer;
 import net.sodiumstudio.dwmg.client.renderer.BefriendedEnderExecutorRenderer;
 import net.sodiumstudio.dwmg.inventory.InventoryMenuBanshee;
@@ -48,7 +52,9 @@ import net.sodiumstudio.dwmg.inventory.InventoryMenuImp;
 import net.sodiumstudio.dwmg.inventory.InventoryMenuKobold;
 import net.sodiumstudio.dwmg.inventory.InventoryMenuNecroticReaper;
 import net.sodiumstudio.dwmg.inventory.InventoryMenuSkeleton;
+import net.sodiumstudio.dwmg.inventory.InventoryMenuSlimeGirl;
 import net.sodiumstudio.dwmg.registries.DwmgEntityTypes;
+import net.sodiumstudio.dwmg.registries.DwmgParticleTypes;
 
 @Mod.EventBusSubscriber(modid = Dwmg.MOD_ID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class DwmgClientSetupEvents 
@@ -76,11 +82,18 @@ public class DwmgClientSetupEvents
         event.registerEntityRenderer(DwmgEntityTypes.HMAG_IMP.get(), ImpRenderer::new);
         event.registerEntityRenderer(DwmgEntityTypes.HMAG_HARPY.get(), HarpyRenderer::new);
         event.registerEntityRenderer(DwmgEntityTypes.HMAG_SNOW_CANINE.get(), SnowCanineRenderer::new);
+        event.registerEntityRenderer(DwmgEntityTypes.HMAG_SLIME_GIRL.get(), SlimeGirlRenderer::new);
         
         event.registerEntityRenderer(DwmgEntityTypes.NECROMANCER_MAGIC_BULLET.get(), MagicBulletRenderer::new); 
         event.registerEntityRenderer(DwmgEntityTypes.BEFRIENDED_GHAST_FIREBALL.get(), c -> new ThrownItemRenderer<>(c, 3.0F, true));
+        event.registerEntityRenderer(DwmgEntityTypes.MAGICAL_GEL_BALL.get(), ThrownItemRenderer::new);
     }
 
+    public static void onRegisterParticleProvider(RegisterParticleProvidersEvent event)
+    {
+    	event.register(DwmgParticleTypes.MAGICAL_GEL_BALL.get(), new MagicalGelBallParticle.Provider());
+    }
+    
 	@SubscribeEvent
 	public static void addGuiScreens(FMLClientSetupEvent event)
 	{
@@ -95,6 +108,7 @@ public class DwmgClientSetupEvents
 		BefriendedGuiScreenMaker.put(InventoryMenuKobold.class, (menu) -> new GuiKobold(menu, menu.playerInventory, menu.mob));
 		BefriendedGuiScreenMaker.put(InventoryMenuImp.class, (menu) -> new GuiImp(menu, menu.playerInventory, menu.mob));
 		BefriendedGuiScreenMaker.put(InventoryMenuFourBaubles.class, (menu) -> new GuiFourBaubles(menu, menu.playerInventory, menu.mob));
+		BefriendedGuiScreenMaker.put(InventoryMenuSlimeGirl.class, (menu) -> new GuiSlimeGirl(menu, menu.playerInventory, menu.mob));
 	}
 	
 }