+++ src/main/java/net/sodiumstudio/dwmg/registries/DwmgBefriendingTypes.java
@@ -21,6 +21,7 @@ import net.sodiumstudio.dwmg.entities.handlers.hmag.HandlerSkeletonGirl;
 import net.sodiumstudio.dwmg.entities.handlers.hmag.HandlerSnowCanine;
 import net.sodiumstudio.dwmg.entities.handlers.hmag.HandlerWitherSkeletonGirl;
 import net.sodiumstudio.dwmg.entities.handlers.hmag.HandlerZombieGirl;
+import net.sodiumstudio.dwmg.entities.handlers.hmag.SlimeGirlBefriendingHandler;
 
 @Mod.EventBusSubscriber(modid = Dwmg.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class DwmgBefriendingTypes {
@@ -93,5 +94,9 @@ public class DwmgBefriendingTypes {
 				ModEntityTypes.SNOW_CANINE.get(),
 				DwmgEntityTypes.HMAG_SNOW_CANINE.get(),
 				new HandlerSnowCanine());
+		BefriendingTypeRegistry.register(
+				ModEntityTypes.SLIME_GIRL.get(),
+				DwmgEntityTypes.HMAG_SLIME_GIRL.get(),
+				new SlimeGirlBefriendingHandler());	// to be changed
 	}
 }